+++ Interface/AddOns/Platynator/Locales.lua
@@ -252,6 +252,8 @@ L["VALUES"] = "Values"
 L["SIGNIFICANT_FIGURES"] = "Significant figures"
 L["ROUNDED"] = "Rounded"
 L["INCLUDE_ELITE_RARES"] = "Include elite rares"
+L["PLAYER_GUILD"] = "Player guild"
+L["NPC_ROLE"] = "NPC role"
 
 L["WHAT_TO_EXPORT"] = "What to export?"
 L["STYLE"] = "Style"