+++ Interface/AddOns/UnhaltedUnitFrames/Core/Config/TagsDatabase.lua
@@ -275,6 +275,30 @@ oUF.Tags.Methods["curpp:abbr"] = function(unit)
     end
 end
 
+oUF.Tags.Methods["curpp:manapercent"] = function(unit)
+    if not unit or not UnitExists(unit) then return "" end
+    local unitPower = UnitPower(unit)
+    local unitPowerType = UnitPowerType(unit)
+    if unitPowerType == Enum.PowerType.Mana and unitPower then
+        local powerPercent = UnitPowerPercent(unit, Enum.PowerType.Mana, true, CurveConstants.ScaleTo100)
+        return string.format("%.f", powerPercent)
+    else
+        return string.format("%s", unitPower)
+    end
+end
+
+oUF.Tags.Methods["curpp:manapercent:abbr"] = function(unit)
+    if not unit or not UnitExists(unit) then return "" end
+    local unitPower = UnitPower(unit)
+    local unitPowerType = UnitPowerType(unit)
+    if unitPowerType == Enum.PowerType.Mana and unitPower then
+        local powerPercent = UnitPowerPercent(unit, Enum.PowerType.Mana, true, CurveConstants.ScaleTo100)
+        return string.format("%.f", powerPercent)
+    else
+        return string.format("%s", AbbreviateValue(unitPower))
+    end
+end
+
 oUF.Tags.Methods["maxpp:abbr"] = function(unit)
     if not unit or not UnitExists(unit) then return "" end
     local unitPowerMax = UnitPowerMax(unit)
@@ -322,6 +346,10 @@ oUF.Tags.Methods["name:colour"] = function(unit)
     local classColourR, classColourG, classColourB = UUF:GetUnitColour(unit)
     local unitName = UnitName(unit) or ""
     return string.format("|cff%02x%02x%02x%s|r", classColourR * 255, classColourG * 255, classColourB * 255, unitName)
+end
+
+oUF.Tags.Methods["resetcolor"] = function(unit)
+    return "|r"
 end
 
 local function ShortenUnitName(unit, maxChars)
@@ -382,6 +410,8 @@ local PowerTags = {
         ["maxpp:colour"] = "Maximum Power with Colour",
         ["maxpp:abbr:colour"] = "Maximum Power with Abbreviation and Colour",
         ["missingpp"] = "Missing Power",
+        ["curpp:manapercent"] = "Current Power but Mana as Percentage",
+        ["curpp:manapercent:abbr"] = "Current Power but Mana as Percentage with Abbreviation",
     },
     {
         "perpp",
@@ -389,6 +419,8 @@ local PowerTags = {
         "curpp:colour",
         "curpp:abbr",
         "curpp:abbr:colour",
+        "curpp:manapercent",
+        "curpp:manapercent:abbr",
         "maxpp",
         "maxpp:abbr",
         "maxpp:colour",
@@ -422,6 +454,7 @@ local MiscTags = {
         ["powercolor"] = "Unit Power Colour - Prefix",
         ["raidcolor"] = "Unit Class Colour - Prefix",
         ["class"] = "Unit Class",
+        ["resetcolor"] = "Resets Colour Prefix",
     },
     {
         "classification",
@@ -432,6 +465,7 @@ local MiscTags = {
         "powercolor",
         "raidcolor",
         "class",
+        "resetcolor",
     }
 }