+++ Interface/AddOns/UnhaltedUnitFrames/Elements/Range.lua
@@ -1,7 +1,11 @@
 local _, UUF = ...
 local isRetail = WOW_PROJECT_ID == WOW_PROJECT_MAINLINE
-local LRC = LibStub("LibRangeCheck-3.0")
+
 UUF.RangeEvtFrames = {}
+
+local IsSpellInRange = C_Spell.IsSpellInRange
+local UnitPhaseReason = UnitPhaseReason
+local CheckInteractDistance = CheckInteractDistance
 
 local RangeEventFrame = CreateFrame("Frame")
 RangeEventFrame:RegisterEvent("PLAYER_TARGET_CHANGED")
@@ -9,81 +13,84 @@ RangeEventFrame:RegisterEvent("UNIT_TARGET")
 RangeEventFrame:RegisterEvent("SPELL_UPDATE_COOLDOWN")
 RangeEventFrame:SetScript("OnEvent", function()
     for _, frameData in ipairs(UUF.RangeEvtFrames) do
-        local frame, unit = frameData.frame, frameData.unit
-        UUF:UpdateRangeAlpha(frame, unit)
+        UUF:UpdateRangeAlpha(frameData.frame, frameData.unit)
     end
 end)
 
-function GetGroupUnit(unit)
-    if strfind(unit, "party") or strfind(unit, "raid") or unit == "player" then
+local function GetGroupUnit(unit)
+    if unit == "player" or unit:match("^party") or unit:match("^raid") then
         return unit
     end
+
     if UnitInParty(unit) or UnitInRaid(unit) then
-        local isInRaid = IsInRaid()
+        local isRaid = IsInRaid()
         for i = 1, GetNumGroupMembers() do
-            local groupUnit = (isInRaid and "raid" or "party")..i
+            local groupUnit = (isRaid and "raid" or "party") .. i
             if UnitIsUnit(unit, groupUnit) then
                 return groupUnit
             end
-        end
-        if not isInRaid and UnitIsUnit(unit, "player") then
-            return "player"
         end
     end
 end
 
 local function IsUnitInRange(unit)
-    local minRange, maxRange = LRC:GetRange(unit, true, true)
-    return (not minRange) or maxRange
+    local spell = UUF.db.profile.General.Range.Spell
+    if spell and IsSpellInRange(spell, unit) ~= nil then
+        return IsSpellInRange(spell, unit)
+    end
+    return CheckInteractDistance(unit, 4)
 end
 
 local function FriendlyIsInRange(realUnit)
     local unit = GetGroupUnit(realUnit) or realUnit
-    if UnitIsPlayer(unit) and (isRetail and UnitPhaseReason(unit) or not isRetail) then
-        return false
+
+    if UnitIsPlayer(unit) then
+        if isRetail then
+            if UnitPhaseReason(unit) then return false end
+        end
     end
-    -- local inRange, checkedRange = UnitInRange(unit)
-    -- if checkedRange and not inRange then return false end
+
     return IsUnitInRange(unit)
 end
 
 function UUF:RegisterRangeFrame(frameName, unit)
-    if not unit or not frameName then return end
-    local unitFrame = type(frameName) == "table" and frameName or _G[frameName]
-    local RangeDB = UUF.db.profile.General.Range
-    table.insert(UUF.RangeEvtFrames, { frame = unitFrame, unit = unit })
-    if RangeDB and RangeDB.Enabled then
-        unitFrame.Range = RangeDB
+    if not frameName or not unit then return end
+
+    local frame = type(frameName) == "table" and frameName or _G[frameName]
+    if not frame then return end
+
+    table.insert(UUF.RangeEvtFrames, { frame = frame, unit = unit })
+
+    if UUF.db.profile.General.Range.Enabled then
+        frame.Range = UUF.db.profile.General.Range
     else
-        unitFrame.Range = nil
+        frame.Range = nil
     end
-    UUF:UpdateRangeAlpha(unitFrame, unit)
+
+    UUF:UpdateRangeAlpha(frame, unit)
 end
 
 function UUF:IsRangeFrameRegistered(unit)
-    for _, frameData in ipairs(UUF.RangeEvtFrames) do
-        if frameData.unit == unit then
-            return true
-        end
+    for _, data in ipairs(UUF.RangeEvtFrames) do
+        if data.unit == unit then return true end
     end
-    return false
 end
 
-function UUF:UpdateRangeAlpha(unitFrame, unit)
-    if not UUF.db.profile.General.Range.Enabled then unitFrame:SetAlpha(1.0) return end
-    if not unitFrame:IsVisible() then return end
-    if not unit or not UnitExists(unit) then return end
-    if unit ~= "player" then
-        local RangeDB = UUF.db.profile.General.Range
-        if not RangeDB then unitFrame:SetAlpha(1.0) return end
-        local inAlpha = RangeDB.InRange or 1.0
-        local outAlpha = RangeDB.OutOfRange or 0.5
-        local unitFrameAlpha
-        if UnitCanAttack('player', unit) or UnitIsUnit(unit, 'pet') then
-            unitFrameAlpha = (IsUnitInRange(unit) and inAlpha) or outAlpha
-        else
-            unitFrameAlpha = (UnitIsConnected(unit) and FriendlyIsInRange(unit) and inAlpha) or outAlpha
-        end
-        unitFrame:SetAlpha(unitFrameAlpha)
+function UUF:UpdateRangeAlpha(frame, unit)
+    local RangeDB = UUF.db.profile.General.Range
+    if not RangeDB or not RangeDB.Enabled then frame:SetAlpha(1) return end
+    if not frame:IsVisible() or not unit or not UnitExists(unit) then return end
+    if unit == "player" then frame:SetAlpha(1) return end
+
+    local inAlpha = RangeDB.InRange or 1
+    local outAlpha = RangeDB.OutOfRange or 0.5
+    local inRange
+
+    if UnitCanAttack("player", unit) or UnitIsUnit(unit, "pet") then
+        inRange = IsUnitInRange(unit)
+    else
+        inRange = UnitIsConnected(unit) and FriendlyIsInRange(unit)
     end
+
+    frame:SetAlpha(inRange and inAlpha or outAlpha)
 end