+++ Interface/AddOns/UnhaltedUnitFrames/CHANGELOG.md
@@ -1,6 +1,6 @@
 # UnhaltedUnitFrames
 
-## [V12.0.10](https://github.com/DaleHuntGB/UnhaltedUnitFrames/tree/V12.0.10) (2026-01-24)
-[Full Changelog](https://github.com/DaleHuntGB/UnhaltedUnitFrames/commits/V12.0.10) 
+## [V12.0.13](https://github.com/DaleHuntGB/UnhaltedUnitFrames/tree/V12.0.13) (2026-01-27)
+[Full Changelog](https://github.com/DaleHuntGB/UnhaltedUnitFrames/commits/V12.0.13) 
 
-- fix tags not updating for bosses  
+- more power tags